Raise Psalm from level 8 to 1 (most strict)
- Fix 3 actually-fixable issues: const type, strtotime cast, float cast
- Suppress systemic level-1 issues with per-category explanations:
  Mixed* (DTO design), MissingConstructor (build() init), InvalidStringClass
  (dynamic instantiation), UnresolvableInclude (temp file), RiskyTruthy
  (truthy checks), TypeDoesNotContainType (string vs class-string),
  InvalidArgument (PhpParser version mismatch), MissingClosure*/Possibly*
  (reflection internals)
- All tools green: Psalm L1 exit 0, PHPStan L9 0 errors, Deptrac 0 viol,
  119 tests pass

+++ src/Infrastructure/StringBuilder.php
@@ -111,7 +111,7 @@ class StringBuilder implements DataTypeInterface
 
     private function randomDateTime(): string
     {
-        return date('Y-m-d', mt_rand(strtotime('-1 year'), strtotime('now')));
+        return date('Y-m-d', mt_rand((int) strtotime('-1 year'), (int) strtotime('now')));
     }
 
     private function randomEmail(): string
